% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vimp_rvimp.R
\name{compare_VIMP_RVIMP}
\alias{compare_VIMP_RVIMP}
\title{Comparison between classical VIMPs and RVIMPs}
\usage{
compare_VIMP_RVIMP(
  formula = NULL,
  ranger.obj = NULL,
  data = NULL,
  residual.model = "Linear",
  ...
)
}
\arguments{
\item{formula}{Object of class formula or character describing the model to fit.}

\item{ranger.obj}{Trained random forest object of class \code{ranger}. If no object is passed \code{formula} is used to generate a new one.}

\item{data}{Training data of class data.frame.}

\item{residual.model}{Type the residuals for each variable shall be based on. Either "Linear" or "RandomForest".}

\item{...}{Additional arguments to the \link{ranger} function.}
}
\value{
Object of class \code{RVIMP_comp} with elements
  \item{\code{residual.model}}{Model the residuals were built with.}
  \item{\code{table}}{Table containing the corresponding values for VIMPs and RVIMPs.}
  \item{\code{plot}}{Plot for visualised presentation of the comparison.}
}
\description{
AUCH HIER SOLLTEN WIR UNS NOCH ETWAS AUSDENKEN
}
\examples{
compare_VIMP_RVIMP(y~.,data=RVIMP_sim_data,residual.model="Linear")
}
\author{
Robert Miltenberger
}
