% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rvimp_sample.R
\name{get_rvimp_sample}
\alias{get_rvimp_sample}
\title{Get sample from RVIMP distribution from RVIMP_test object}
\usage{
get_rvimp_sample(obj, which = "all")
}
\arguments{
\item{obj}{\code{RVIMP_test} object.}

\item{which}{Variables the distribution shall be returned for.}
}
\value{
Matrix with samples from the RVIMP distribution the test procedure was performed with for desired variables
}
\description{
Get sample from RVIMP distribution from RVIMP_test object
}
\author{
Robert Miltenberger
}
