% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_test.R
\name{multiple_test.RVIMP}
\alias{multiple_test.RVIMP}
\title{Adjustment for multiple test procedure}
\usage{
multiple_test.RVIMP(test.obj = NULL, method = "Bonferroni")
}
\arguments{
\item{test.obj}{Object of class test_RVIMP.}

\item{method}{Which procedure shall be applied for adjustment? Either "\code{Bonferoni}", "\code{Sidak}" or "\code{Holm}".}
}
\value{
Object of class \code{RVIMP_multiple_test} with elements
 \item{\code{method}}{With multiple test procedure was used.}
 \item{\code{confounded}}{Name of the variables tested for dependencies.}
 \item{\code{multiple_tests}}{List with results for multiple test procedure for desired variables \code{confounded}.}
}
\description{
Adjustment for multiple test procedure
}
\examples{
RVIMP_test_object<-test.RVIMP(y~.,data=RVIMP_sim_data,Z=c("x1","x2"))
multiple_test.RVIMP(RVIMP_test_object,method="Bonferroni")
}
