% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RVIMP.R
\name{plot.RVIMP}
\alias{plot.RVIMP}
\title{Plot the results of a RVIMP object}
\usage{
\method{plot}{RVIMP}(x, which = "all", ask = T, ...)
}
\arguments{
\item{x}{\code{RVIMP} object.}

\item{which}{For which variable shall the plot be generated. Either a \link{vector} with the names of the desired variables or \code{"all"} to generate the plot for each variable the RVIMPs were calculated for.}

\item{ask}{Logical. If \code{TRUE} the user is asked to press Enter before a new figure is drawn. If \code{FALSE} all figures are drawn at once.}

\item{...}{Further arguments to the \link{plot} function.}
}
\description{
Generates the plot to compare VIMPs from model_A and model_B with for every variable checked for dependencies.
}
