% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp.R
\name{plot.RVIMP_comp}
\alias{plot.RVIMP_comp}
\title{Print results for the comparison between VIMPs and RVIMPs}
\usage{
\method{plot}{RVIMP_comp}(x, ...)
}
\arguments{
\item{x}{\code{RVIMP_comp} object.}

\item{...}{Further argumentss to the \link{print} function.}
}
\description{
Print results for the comparison between VIMPs and RVIMPs
}
