% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_RVIMP.R
\name{print.RVIMP}
\alias{print.RVIMP}
\title{Prints the main results from a RVIMP object}
\usage{
\method{print}{RVIMP}(x, ...)
}
\arguments{
\item{x}{\code{RVIMP} object.}

\item{...}{Further arguments to the \link{print} function.}
}
\description{
Summarizes the main results from a RVIMP object
}
