% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_comp.R
\name{print.RVIMP_comp}
\alias{print.RVIMP_comp}
\title{Print results for comparison between VIMPs and RVIMPs}
\usage{
\method{print}{RVIMP_comp}(x, ...)
}
\arguments{
\item{x}{\code{RVIMP_comp} object.}

\item{...}{Further arguments to the \link{print} function.}
}
\description{
Summarizes the results of the comparison between VIMPs and RVIMPs
}
