% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_RVIMP_multiple_test.R
\name{print.RVIMP_multiple_test}
\alias{print.RVIMP_multiple_test}
\title{Prints the main results from a RVIMP_multiple_test object}
\usage{
\method{print}{RVIMP_multiple_test}(x, ...)
}
\arguments{
\item{x}{\code{RVIMP_multiple_test} object.}

\item{...}{Further arguments to the \link{print} function.}
}
\description{
Summarizes the main results from a RVIMP_test object
}
