% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_markdown.R
\name{run_markdown}
\alias{run_markdown}
\title{Render R Markdown to analyze a data set for dependencies using RVIMP package}
\usage{
run_markdown(
  formula = NULL,
  data = NULL,
  Z = "all",
  residual.model = "Linear",
  seed = 42,
  reps = 100,
  sample.seed = 50,
  alpha.one.sided = 0.05,
  num.trees = 200,
  multi.test = T,
  procedure = "Bonferroni",
  file.name = NULL,
  file.directory = NULL
)
}
\arguments{
\item{formula}{Object of class formula or character describing the model to fit.}

\item{data}{Training data of class data.frame.}

\item{Z}{Name of all variables the RVIMPs shall be calculated for. If \code{Z="all"} the procedure is done for all independed variables from \code{formula}.}

\item{residual.model}{Modeltype the residuals of \code{Z} shall be based on. Either "Linear" or "RandomForest".}

\item{seed}{Seed passed to ranger function for generating the Random Forest. Must be set for reproducible results. Note: \code{seed} is passed to the \code{ranger} function to generate the Random Forest, whereas \code{sample.seed} is needed to draw the resamplings for the test procedure.}

\item{reps}{Number of respamplings to estimate the RVIMP distribution with \link{get_distribution}.}

\item{sample.seed}{Seed to use for \link{get_distribution}. Must be set for reproducible results.}

\item{alpha.one.sided}{Type I error probability.}

\item{num.trees}{Number of trees within each ranger object.}

\item{multi.test}{Logical. Shall a multiple test procedure be performed?}

\item{procedure}{If \code{multi.test=T} either \code{"Bonferroni"}, \code{"Holm"}, or \code{"Sidak"}.}

\item{file.name}{The name of the output file. If using \code{NULL} then the output file is named "report_RVIMP.html".}

\item{file.directory}{The output directory for the rendered \code{file.name}. This allows for a choice of an alternate directory to which the output file should be written (the default output directory is within the package folder). If a path is provided with a filename in \code{file.name} the directory specified here will take precedence. Please note that any directory path provided will create any necessary directories if they do not exist.}
}
\description{
Generates a detailed analysis to check for dependencies within a data set using Permutation Variable Importance Measures. Provides results as a .html-file in the desired directory. All intermediate files (e.g. generated plots) are written to the same directory.
}
