% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_RVIMP_multiple_test.R
\name{summary.RVIMP_multiple_test}
\alias{summary.RVIMP_multiple_test}
\title{Summary of the results for a RVIMP_multiple_test for dependencies}
\usage{
\method{summary}{RVIMP_multiple_test}(object, ...)
}
\arguments{
\item{object}{\code{RVIMP_multiple_test} object.}

\item{...}{Further arguments to the \link{summary} function.}
}
\description{
Summary of the results for a RVIMP_multiple_test for dependencies
}
\author{
Robert Miltenberger
}
