% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.RVIMP}
\alias{test.RVIMP}
\title{Test procedure to check for dependencies within covariates based on Permutation Variable Importance Measures in Random Forests}
\usage{
test.RVIMP(
  formula = NULL,
  ranger.obj = NULL,
  data = NULL,
  Z = "all",
  residual.model = "Linear",
  residuals = NULL,
  alpha.one.sided = 0.05,
  keep.ranger = T,
  reps = 100,
  sample.seed = NULL,
  seed = NULL,
  num.trees = 500,
  ...
)
}
\arguments{
\item{formula}{Object of class formula or character describing the model to fit.}

\item{ranger.obj}{Trained random forest object of class \link{ranger}. If no object is passed \code{formula} is required to generate a new one. Passed \code{ranger} object must contain \code{variable.importance}.}

\item{data}{Training data of class data.frame.}

\item{Z}{Name of all variables that shall be checked for dependencies. If \code{Z="all"} the procedure is done for all independed variables from \code{formula}.}

\item{residual.model}{Model the residuals of \code{Z} shall be estimated with. Either \code{"Linear"} or \code{"RandomForest"} if residuals shall be estimated within the algorithm or \code{"Precalculated"} if precalculated residuals are passed to \code{residuals} by the user.}

\item{residuals}{Residuals, if they are precalculated.}

\item{alpha.one.sided}{Type I error probability.}

\item{keep.ranger}{Logical. Shall either the passed \code{ranger.obj} or the newly generated one be part of the result.}

\item{reps}{Number of respamplings to estimate the RVIMP distribution on. Each resampling is based on a data set with a sample size of 0.632 compared to the original data set.}

\item{sample.seed}{Seed to draw the \code{reps} resamplings of the original data set. Must be set for reproducible results.}

\item{seed}{Seed passed to ranger function for generating the Random Forest. Must be set for reproducible results. Note: \code{seed} is passed to the \code{ranger} function to generate the Random Forest, whereas \code{sample.seed} is needed to draw the resamplings for the test procedure.}

\item{num.trees}{Number of trees within each ranger object.}

\item{...}{Further arguments to the \link{ranger} function.}
}
\value{
Object of class \code{RVIMP_test} with elements
  \item{\code{call}}{Function call.}
  \item{\code{variables}}{Name of the variables tested for dependencies.}
  \item{\code{reps}}{Number of resamplings to approximate the RVIMP distribution.}
  \item{\code{alpha}}{Desired type I error rate.}
  \item{\code{residual.model}}{Model the residuals were bulit with.}
  \item{\code{ranger}}{Object of class \code{ranger} for the original model or \code{NULL} if \code{keep.ranger=F}.}
  \item{\code{RVIMPs}}{List of test results for each tested variable.}
  \item{\code{VIMP}}{Permutation Variable Importance of the original model.}
}
\description{
Performes a resampling test to check if a variables importance is at least partially induced from other variables within a Random Forest.
}
\examples{
test.RVIMP(y~.,data=RVIMP_sim_data,Z="x1")
}
\seealso{
\code{\link{RVIMP}} \code{\link{multiple_test.RVIMP}}
}
\author{
Robert Miltenberger
}
