% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVIMP.R
\name{RVIMP}
\alias{RVIMP}
\title{Estimation of permutation variable importance measures corrected for dependencies in Random Forests}
\usage{
RVIMP(
  formula = NULL,
  ranger.obj = NULL,
  data = NULL,
  Z = "all",
  residual.model = "Linear",
  keep.ranger = T,
  ...
)
}
\arguments{
\item{formula}{Object of class formula or character describing the model to fit.}

\item{ranger.obj}{Trained random forest object of class \link{ranger}. If no object is passed \code{formula} is used to generate a new one. Passed \code{ranger} object must contain \code{variable.importance}.}

\item{data}{Training data of class data.frame.}

\item{Z}{Name of all variables the RVIMPs shall be calculated for. If \code{Z="all"} the procedure is done for all independed variables from \code{formula}.}

\item{residual.model}{Modeltype the residuals of \code{Z} shall be based on. Either \code{"Linear"} or \code{"RandomForest"}.}

\item{keep.ranger}{Shall generated ranger object be kept.}

\item{...}{Further arguments to the \link{ranger} function. For reproducible results \code{seed} must be set.}
}
\value{
Object of class \code{RVIMP} with elements
  \item{\code{call}}{Function call.}
  \item{\code{variables}}{Name of the variables the RVIMPs are calculated for.}
  \item{\code{residual.model}}{Model the residuals were bulit with.}
  \item{\code{ranger}}{Object of class \code{ranger} for the original model or \code{NULL} if \code{keep.ranger=F}.}
  \item{\code{RVIMPs}}{List of results for each variable the RVIMP shall be calculated for.}
  \item{\code{VIMP}}{Permutation Variable Importance of the original model.}
}
\description{
RVIMP is a way to improve the interpretation of variables in Random Forests. The procedure is based on the permutation variable importance measure, first mentioned by Breiman 2001. HIER SOLLTEN WIR UNS DANN VERMUTLICH ZUSAMMEN ETWAS AUSDENKEN.
}
\examples{
RVIMP(formula=y~.,data=RVIMP_sim_data,Z = "all",residual.model = "Linear")
}
\references{
\itemize{
  \item Breimann L. Random Forests, Mach. Learn. 2001; 45(1):5-32.
  \item Gregorutti B, Saint-Pierre P. Correlation and variable importance in random forests, Stat Comput 2017; 27:659-678
  \item Wright MN, Ziegler A. ranger: A fast implementation of random forests for high dimensional data in C++ and R. J Stat Softw 2017, 77:1-17. \doi{https://doi.org/10.18637/jss.v077.i01}
 }
}
\seealso{
\code{\link{test.RVIMP}}
}
